Update Swift+UIKit example to 3.2.0
