+++ ios/Swift+UIKit/IntegrationDemo_Swift_UIKit/SceneDelegate.swift
@@ -11,11 +11,23 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
 	var window: UIWindow?
 
+	let apiToken: String = "0b5dcbae8151c1b82d69697dce004bf2"
+	let userIdentifier: String = "public-demo-user"
+
+	let tapDelegates: TapResearchDelegates = TapResearchDelegates()
+
 	func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
 		// Use this method to optionally configure and attach the UIWindow `window` to the provided UIWindowScene `scene`.
 		// If using a storyboard, the `window` property will automatically be initialized and attached to the scene.
 		// This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
 		guard let _ = (scene as? UIWindowScene) else { return }
+
+		// Initialize TapResearch
+		TapResearchSDK.initialize(withAPIToken: apiToken, userIdentifier: userIdentifier, sdkDelegate: tapDelegates) { (error: NSError?) in
+			if let e = error {
+				print(e.localizedDescription as Any)
+			}
+		}
 	}
 
 	func sceneDidDisconnect(_ scene: UIScene) {
@@ -45,5 +57,5 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 		// Use this method to save data, release shared resources, and store enough scene-specific state information
 		// to restore the scene back to its current state.
 	}
-
+	
 }
