+++ ios/Swift+UIKit/IntegrationDemo_Swift_UIKit/ViewController.swift
@@ -8,27 +8,19 @@
 import UIKit
 import TapResearchSDK
 
-///---------------------------------------------------------------------------------------------
-///---------------------------------------------------------------------------------------------
-class ViewController : UIViewController,
-					   UITextFieldDelegate,
-					   UITableViewDelegate,
-					   UITableViewDataSource,
-					   TapResearchContentDelegate
-{
+class ViewController : UIViewController, UITextFieldDelegate, UITableViewDelegate, UITableViewDataSource, TapResearchContentDelegate {
 
 	@IBOutlet weak var tableView: UITableView!
 	@IBOutlet weak var textField: UITextField!
 	@IBOutlet weak var placementStatus: UILabel!
 
 	var knownPlacements: [String] = [
-        "default-placement",
-        "interstitial-placement",
-        "floating-interstitial-placement",
-        "banner-placement"
+		"default-placement",
+		"interstitial-placement",
+		"banner-placemenet",
+		"floating-interstitial-placement"
 	]
 
-	///---------------------------------------------------------------------------------------------
 	override func viewDidLoad() {
 		super.viewDidLoad()
 
@@ -38,7 +30,6 @@ class ViewController : UIViewController,
 
 	//MARK: - UITextFieldDelegate
 
-	///---------------------------------------------------------------------------------------------
 	func textFieldDidChangeSelection(_ textField: UITextField) {
 
 		if let _ = placementStatus.text {
@@ -46,7 +37,6 @@ class ViewController : UIViewController,
 		}
 	}
 
-	///---------------------------------------------------------------------------------------------
 	func textFieldShouldReturn(_ textField: UITextField) -> Bool {
 
 		showPlacement()
@@ -55,14 +45,12 @@ class ViewController : UIViewController,
 
 	//MARK: - Actions and button handlers
 
-	///---------------------------------------------------------------------------------------------
 	@IBAction func showPlacement() {
 		guard let text: String = textField.text else { return }
 
 		if TapResearchSDK.canShowContent(forPlacement: text) {
-            let customParameters = ["param1": 123, "param2": "abc"] as [String : Any]
-
-			TapResearchSDK.showContent(forPlacement: text, delegate: self, customParameters: customParameters) { (error: NSError?) in
+			//let customParameters = ["param1": 123, "param2": "abc"] as [String : Any]
+			TapResearchSDK.showContent(forPlacement: text, delegate: self/*, customParameters: customParameters*/) { (error: NSError?) in
 				if let error = error {
 					self.placementStatus.text = "\(error.code) \(error.localizedDescription)"
 				}
@@ -83,7 +71,6 @@ class ViewController : UIViewController,
 
 	//MARK: - Table delegate and datasource
 	
-	///---------------------------------------------------------------------------------------------
 	func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
 		
 		self.tableView.deselectRow(at: indexPath, animated: true)
@@ -100,29 +87,24 @@ class ViewController : UIViewController,
 		}
 	}
 	
-	//---------------------------------------------------------------------------------------------
 	func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
 		return knownPlacements.count
 	}
 
-	///---------------------------------------------------------------------------------------------
 	func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
 		return "Known Placements"
 	}
 	
-	///---------------------------------------------------------------------------------------------
 	func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
 		return PlacementCell.cell(tableView: tableView, placement: knownPlacements[indexPath.row])
 	}
 
 	//MARK: - TapResearchContentDelegate
 
-	///---------------------------------------------------------------------------------------------
 	func onTapResearchContentShown(forPlacement placement: String) {
 		print("onTapResearchContentShown(\(placement))")
 	}
 	
-	///---------------------------------------------------------------------------------------------
 	func onTapResearchContentDismissed(forPlacement placement: String) {
 		print("onTapResearchContentDismissed(\(placement))")
 	}
