+++ ios/Swift+UIKit/Podfile
@@ -1,11 +1,11 @@
 # Uncomment the next line to define a global platform for your project
-platform :ios, '13.0'
+platform :ios, '12.0'
 
 target 'IntegrationDemo_Swift_UIKit' do
   # Comment the next line if you don't want to use dynamic frameworks
   use_frameworks!
 
   # Pods for swiftui-example-tapresearch
-  pod 'TapResearch', '3.1.0-beta1'
+  pod 'TapResearch', '3.2.0-beta1-15'
 
 end
